airflow/requirements.txt
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
beautifulsoup4
pandas
requests
python-jobspy
